+++ pypsa/components/_types/carriers.py
@@ -1,0 +1,26 @@
+"""Carriers components module."""
+
+from __future__ import annotations
+
+from pypsa.components.components import Components
+
+
+class Carriers(Components):
+    """
+    Carriers components class.
+
+    This class is used for carrier components. All functionality specific to
+    carriers is implemented here. Functionality for all components is implemented in
+    the abstract base class.
+
+    .. warning::
+        This class is under ongoing development and will be subject to changes.
+        It is not recommended to use this class outside of PyPSA.
+
+    See Also
+    --------
+    pypsa.components.abstract.Components : Base class for all components.
+
+    """
+
+    pass
